Improves frontend and backend setup
Streamlines the frontend Dockerfile by using `package*.json` to copy both package.json and package-lock.json.
Adds missing size attribute to buttons on the auth forms.
Fixes an issue where focus was not properly set on the name field of the update class form.
Adds README.MD with instructions for running the project.